Building an RAG-based Chatbot using LLM and VectorDB/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
tiktoken
openai
